requirements.txt
# Core ML frameworks
torch>=2.0.0
torchvision>=0.15.0
tensorflow>=2.13.0
transformers>=4.35.0
ultralytics>=8.0.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
opencv-python>=4.8.0
Pillow>=10.0.0
albumentations>=1.3.0

# Model export/optimization
onnx>=1.14.0
onnxruntime>=1.16.0
tensorflowjs>=4.10.0

# Training utilities
tensorboard>=2.13.0
wandb>=0.15.0
mlflow>=2.7.0
optuna>=3.3.0
ray[tune]>=2.7.0

# Deployment
gradio>=4.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
streamlit>=1.25.0
huggingface-hub>=0.20.0
runpod>=1.6.0

# Development
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0
ipykernel>=6.25.0
jupyter>=1.0.0

# Utilities
pyyaml>=6.0.1
python-dotenv>=1.0.0
tqdm>=4.65.0
click>=8.1.0
rich>=13.5.0
typer>=0.9.0